Modified update api and use Global exception handling

+++ src/main/java/com/codingshuttle/youtube/LearningRESTAPIs/repository/StudentRepository.java
@@ -3,5 +3,5 @@ package com.codingshuttle.youtube.LearningRESTAPIs.repository;
 import com.codingshuttle.youtube.LearningRESTAPIs.entity.Student;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StudentRepository extends JpaRepository<Student, Long> {
+public interface StudentRepository extends JpaRepository<Student, Integer> {
 }
